Remove the action key from history entries
The action is save in the entries' parent dict key ('opened' or
'closed') so there is no need for the action to appear again in every
entry.

There were cases previously were the action was missing I suppose (for
some reason) and 'accessed' was used, but that shouldn't happen anymore.

Closes #30.