Track and restore selected index after delete
Since the quick panel now uses a separate copy of the history, the logic
to track the currently selected index and restore it after deleting a
history entry was broken, since the index in the quick panel list will
no longer be consistent with the index in the latest history list. This
change now moves the selected index tracking into the OpenRecentlyClosedFile
command object itself, and orchestrates history entry deletion through it.
This introduces a breaking change in the arguments accespted by the
command.