Fix race between panel close and new command invocation
On running the delete_file_history_entry command, the currently open
quick panel for the open_recently_closed_file command is closed and
the command is re-invoked. However, the closing of the overlay will
cause the on_done() callback of the quick panel to get invoked, which will
race with the run() of the new command invocation. The on_done() callback
clears the cached history list, while the new command invocation refreshes
it. If the on_done() gets invoked after the new command invocation has run,
this will cause the list to get cleared while the quick panel is still open,
causing any future operations on the quick panel list to fail. The fix is
to orchestrate the closing and re-opening of the quick panel through the
open_recently_closed_file command itself, so that everything is
synchronized correctly.