Properly handle files that exist in other groups
Involves some rewriting of general preview logic, opening and closing,
while also using the `sublime.FORCE_GROUP` flag, if it is available.
Files will now always be previewed in the current group and the focus
is switched to the other group if it "really" opened (quick or normal)
and already exists. For quick opening, this requires reopening the panel
because of a ST bug that auto-closes the panel if group focus changes.

In the process, more commands have been eliminated and migrated into the
"open_recently_closed_file" command, because it stores the state.

Also introduce "reopen_file_in_current_group" setting that will
force-open a file in the current group even if it exists in another
already.

Fixes #23.