Create a copy of the history for use in the panel
The quick panel now uses a copy of the history instead of directly using
the underlying history data structure. This avoids problems that arise
due to the history changing while the quick panel is open, which leaves
the quick panel in an inconsistent state and can cause an incorrect file
to be opened in some cases. Since the quick panel now uses a copy, the
"queued delete" functionality is no longer required.

Cherry-picked and modified from @rahul-ramadas.